Move Auth to shared folder
