+++ common/auth/types.ts
@@ -69,3 +69,45 @@ export const OrgCategory = Union(
   L("Other")
 )
 export type OrgCategory = Static<typeof OrgCategory>
+
+export const OrgCategories = [
+  "Agriculture",
+  "Animal Rights",
+  "Automotive",
+  "Civil Rights",
+  "Communications",
+  "Consumer Goods",
+  "Criminal Justice",
+  "Prison Reform",
+  "Education",
+  "Elder Care",
+  "Employment/Labor",
+  "Environment",
+  "Financial",
+  "Food",
+  "Good Government",
+  "Human Rights",
+  "Children's Rights",
+  "Death Penalty",
+  "Disabeled Rights",
+  "LGBTQ+ Rights",
+  "Housing",
+  "Immigration",
+  "Insurance",
+  "Internet & Technology",
+  "Legal",
+  "Medical/Health",
+  "Poverty",
+  "Privacy",
+  "Racial Justice",
+  "Regional",
+  "Refugee",
+  "Reproductive Health",
+  "Pharmaceuticals",
+  "Small & Local Business",
+  "Taxes",
+  "Water",
+  "Women's Rights",
+  "Multi-issue",
+  "Other"
+]
