+++ components/EditProfilePage/PersonalInfoTab.tsx
@@ -5,7 +5,7 @@ import { ProfileHook } from "../db"
 import Input from "../forms/Input"
 import { TitledSectionCard } from "../shared"
 import { YourLegislators } from "./YourLegislators"
-import { OrgCategory, OrgCategories } from "components/auth"
+import { OrgCategory, OrgCategories } from "common/auth/types"
 import { TooltipButton } from "components/buttons"
 import { useTranslation } from "next-i18next"
 import styled from "styled-components"
