+++ functions/src/notifications/helpers.ts
@@ -1,5 +1,5 @@
 import { Timestamp } from "../../../common/types"
-import { Frequency } from "../auth/types"
+import { Frequency } from "../../../common/auth/types"
 import {
   startOfDay,
   addMonths,
