+++ tests/integration/auth.test.ts
@@ -1,6 +1,6 @@
 import { UserRecord } from "firebase-admin/auth"
 import { nanoid } from "nanoid"
-import { Role } from "../../components/auth"
+import { Role } from "common/auth/types"
 import { setRole } from "../../functions/src/auth"
 import { terminateFirebase, testAuth, testDb } from "../testUtils"
 import { getProfile } from "./common"
